requirements.txt
absl-py==0.12.0
addict==2.4.0
aiohttp==3.7.4.post0
albumentations==1.0.0
antlr4-python3-runtime==4.8
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
att-grid-generator==0.0.0
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
block-extractor-cuda==0.0.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
chumpy==0.70
clang==5.0
click==8.0.1
config==0.5.0.post0
configer==1.4.1
configparser==5.0.2
ctcdecode==1.0.2
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
defusedxml==0.7.1
dominate==2.6.0
einops==0.3.0
entrypoints==0.3
fire==0.4.0
flatbuffers==1.12
flow-vis==0.1
freetype-py==2.2.0
fsspec==2021.6.1
future==0.18.2
fvcore==0.1.2.post20210113
gast==0.4.0
gensim==3.8.3
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.41.0
h5py==3.1.0
human-body-prior==0.9.3.0
hydra-core==1.0.6
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.4
imgaug==0.4.0
importlib-metadata==3.7.3
importlib-resources==5.1.3
iopath==0.1.2
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonpatch==1.32
jsonpointer==2.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
llvmlite==0.36.0
lmdb==1.2.1
local-attn-reshape-cuda==0.0.0
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
moviepy==2.0.0.dev2
multidict==5.1.0
natsort==7.1.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.5
neural-renderer==1.1.3
nlg-eval==2.3
nltk==3.6.2
notebook==6.2.0
numba==0.53.1
numpy==1.20.3
oauthlib==3.1.0
omegaconf==2.0.6
open3d==0.12.0
open3d-python==0.7.0.0
opencv-python==4.5.1.48
opencv-python-headless==4.5.2.54
opendr==0.77
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.0
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
plyfile==0.7.3
portalocker==2.0.0
proglog==0.1.9
prometheus-client==0.9.0
prompt-toolkit==3.0.17
protobuf==3.16.0
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
pyDeprecate==0.3.0
pyglet==1.5.18
Pygments==2.8.1
PyOpenGL
PyOpenGL-accelerate==3.1.5
pyparsing==2.4.7
pyrender==0.1.45
pyrsistent==0.17.3
python-dateutil==2.8.1
pytorch-lightning==1.3.7.post0
pytorch-msssim==0.2.1
pytorch3d==0.3.0
pytz==2020.5
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==22.0.3
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
resample2d-cuda==0.0.0
rsa==4.7.2
sacrebleu==1.5.1
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.0
Send2Trash==1.5.0
Shapely==1.7.1
simplejson==3.17.2
six==1.15.0
sklearn==0.0
smart-open==5.1.0
smplx==0.1.28
tabulate==0.8.7
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
tensorflow==2.6.0
tensorflow-estimator==2.6.0
tensorflow-gpu==2.6.0
termcolor==1.1.0
terminado==0.9.3
test-tube==0.7.5
testpath==0.4.4
Theano==1.0.5
threadpoolctl==2.1.0
tifffile==2021.1.11
torch==1.8.0
torchfile==0.1.0
torchgeometry==0.1.2
torchmetrics==0.3.2
torchstat==0.0.7
torchvision==0.9.0
tornado==6.1
tqdm==4.56.0
traitlets==5.0.5
transforms3d==0.3.1
trimesh==3.9.23
typing-extensions==3.7.4.3
urllib3==1.26.4
visdom==0.1.8.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xdg==5.1.0
yacs==0.1.8
yarl==1.6.3
zipp==3.4.1
